Adding new package trace-cmd v2.9 for photon
trace-cmd build requires certain libraries during
runtime that it places in the /usr/lib/traceevent/plugins
directory on installation.
perf also uses the same libraries from the same directory.
As a result, this caused conflicts during installation of
the second package if either trace-cmd or linux-tools
(that includes perf) was already installed.

To address the above problem, the shared libraries are
are generated as a subpackage of trace-cmd
called traceevent-plugins. Both linux-tools and trace-cmd
require traceevent-plugins to be able deparse
certain events correctly. Steven Rostedt, the author
and maintainer of trace-cmd has confirmed that the
libraries for both are kept in sync and that it is
safe to share the libraries between the two.

Change-Id: I8acb00c5917dddedab19429d71830795786d34c3
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/10590
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>