Scheduler: Fix makecheck getting stuck
Build gets stuck when running rpmcheck. This happened after change
d8eaa7441f5("Update Scheduler to use graph while building")
What happens is when control goes to '_getListNextPackagesReadyToBuild()'
it again checks whether pkg is in 'listOfAlreadyBuiltPackages', which in
case of rpmcheck is true. So it shouldn't matter if package has been already
built. Made following changes to fix the issue

Removed check for pkg in listOfAlreadyBuiltPackages when getting
next pkg to build, because if pkg is in listOfPackagesToBuild
then it will not be in already built unless we are running a makecheck
in which case we dont care as we assume build has already happened.

If constants.rpmcheck is true in that case we put that package in
'listOfPackagesNextToBuild' directly, we need not check if all children of
the pkg are built or not. Makecheck being an independent job.

We dont need to markPkgNodeAsBuilt because when rpmcheck is true we
dont generate graph. Hence mapPackagesToGraphNodes is empty.

Changed python2 packages from constants.py to use python3 since it was
throwing error that package not available after python2 was removed.

Change-Id: I7582e79d9eae239ba4f6fc2de77a49bd1d54de5a
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11187
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>