imagegenerator: fixed azure image  not building inside docker container
When trying to build cloud images inside docker container
azure fails because for vhd we run qemu command inside docker
resulting in docker inside docker(dind).

Docker inside docker is not a recommended option  due to various
issues. Exposing docker socket to run dind is not really a
proper nesting instead it creates a sibling container.
Since in our scenario we are  mounting volume when running
docker command. This means mount path of build should always be same
as that of inside docker.
There is another docker dind image which creates proper encapsulation
but we dont want that.

So its better we check whether our environment is docker or not.
If yes then we only run the command if not we run command
inside photon docker.
To check if we are inside docker environment we check
/proc/self/cgroup. Inside docker this return with field
as docker 'name=systemd:/docker/0494....' which we check to
verify if we are inside docker or not

Change-Id: Ifd6e422aa6246de9e4b67a343f9cb2998bdf2510
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11205
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ankit Jain <ankitja@vmware.com>