linux-esx: 9p: Don't use writeback fid for cache when enabled for VDFS
Writeback fid is used for data caching. It is utilized to write dirty
pages back to the server.Writeback fid is associated with an inode. An
inode may still be cached even in case file is closed,the writeback_fid
associated to the inode in cache is still opened, and the file
descriptor on host of the writeback_fid is also kept opened.This can
cuase too many files opened at the host.VDFS does not use caching for
data and to avoid above issue, usage of writeback fid for VDFS has been
prohibited.

Change-Id: I433ff6f0eeeb07e9186fc4069cd2be459e8bab68
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11324
Reviewed-by: Kevin Kong <kkong@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ajay Kaher <akaher@vmware.com>