kernels: Fix for CVE-2019-19377
  -  In BTRFS file system, there is a chance of usage after
free if dirty pages or dirty tree block would be present.
Hence prevent the metadata write bio to be submitted to
the btrfs work queue if FS already has an error.

  Modified the patch such that it is applicable to 4.4
stable branch. Tested with the scripts to mount a corrupt
BTRFS image and verified that there is no use-after-free
issue.

Change-Id: If0410ba9b1e855084383a64ac41ff7dbeb174cb2
Signed-off-by: srinidhira0 <srinidhir@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11297
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>