linux-esx: 9p: Fix recovery logic and cleanup tags when unmounting
We maintain a list of recovery descriptors to track the operations that need to
recover fids or locks in case of server crash. However, there is a gap between
when the op completes and when it is added to the recovery list. If server
restarts during this time, then the recovery thread will immediately start.

Because some of the operations have completed but not yet added the descriptor
in the recovery list, we may miss some recovery ops which will result in
disruptions when those fid or locks are used by next set of file system
operations.

Adding a new reference count for the recoverable inflight requests to keep track
of the requests which are inflight but not yet added to recovery list. The
recovery thread will always wait for the refcount to be zero before starting

For the claimtags request we keep a dummy request in the tag tree which should
be ignored when cleaning up the tags before unmount. Removing all the dummy tags
as part of tag cleanup before processing the valid in-use tags.

Change-Id: I581fbbe5121bb6420e763109b273b7374db5979a
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11177
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>