dracut: fix hostonly flag setting logic to generate proper initrd
From the following commit:
https://github.com/dracutdevs/dracut/commit/f4ad2e5c8964b6c7b50420299a0096bade6a40bd

Kernel modules won't be added to initrd image if hostonly flag is set.
As a result when we generate cloud images in Photon(Azure for example)
and cloud image won't boot.

From this commit:
https://github.com/dracutdevs/dracut/commit/32f68c1f9ac3720e8ce4b95a09c0ce680d5da786

dracut asks for root password if hostonly mode is set, so while
generating initrd for a different host, if dracut fails for some reason
it'll drop to an emergency shell and it won't accept any password as we
are disabling root login in our cloud images.

This patch makes it right by setting hostonly to true only if we are NOT
generating initrd image inside docker environament.

Further observations:
If hostonly mode is set:
initrd.img = ~8MB

If hostonly mode is unset:
initrd.img = ~25MB

Read more  about hostonly at:
https://fedoraproject.org/wiki/Features/DracutHostOnly#Detailed_Description
https://man7.org/linux/man-pages/man5/dracut.conf.5.html

Change-Id: I62f742a118d9af5bc93f250eb027921213bf77d5
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11448
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>
Reviewed-by: Susant Sahani <ssahani@vmware.com>