Architecture dependent package installation
This involves using an extra key apart from 'packages'
such as 'packages_x86_64' and maintaining architecture
dependent packages under that key.

This solves issue of package installation failure when
installed on different architecture.

For ex - 'grub2-pc' is built for x86_64 and not for aarch64.
So, while installing developer version in arm grub2-pc should
not be included in list of packages to install.

Change-Id: Ia277dcd6f5afbfbb4c479ff69fb09a581cb22cfa
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11520
Reviewed-by: Anish Swaminathan <anishs@vmware.com>
Tested-by: Anish Swaminathan <anishs@vmware.com>