Realtime: Skip flavor selection for aarch64
linux-esx: Moved linux-esx to packages_x86_64 to skip
	   linux selection screen during iso installation
	   for aarch64.

Architecture dependent package installation

This involves using an extra key apart from 'packages'
such as 'packages_x86_64' and maintaining architecture
dependent packages under that key.

This solves issue of package installation failure when
installed on different architecture.

For ex - 'grub2-pc' is built for x86_64 and not for aarch64.
So, while installing developer version in arm grub2-pc should
not be included in list of packages to install.

Change-Id: I95714e8a4a1b85c4eadffbb4d07f1c139495a9ec
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11838
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Anish Swaminathan <anishs@vmware.com>