libsolv: Further extend choicerule filtering check
When we install tuned, instead of updating to latest
linux version libsolv was forcing to install linux-aws
package.

There is actually a countermeasure in the code
(the so-called "choice rules") to prevent the
installation of new packages if a dependency can also be
met by updating a package, but for some reason it
does not work in the case we install tuned.

This commit will fix that issue.

Rebuilt tdnf with latest libsolv patches.

Change-Id: I9fa61eb4f78df7d537ac80a351edba1063c8a07c
Signed-off-by: Tapas Kundu <tkundu@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11744
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Keerthana K <keerthanak@vmware.com>