[9P] fscache: Only fetch attr from inode cache when cache is valid
When setattr is called, the ctime of the file system entry will be updated.
The ctime in the inode cache will be inconsistent with the server. atime and
mtime is incorrect as well as they're truncated according to time granularity.
See setattr_copy().
In this case, attr cache in the inode has already been marked as invalid
by v9fs_invalidate_inode_attr().
When a following getattr is called, we only fetch from attr cache when
cache in inode is valid. If it invalid, the attr will be fetched from the
server, this ensures 9p client always get consistent timestamp.
After we fetch attr from 9p server, the cache will become valid by calling
v9fs_stat2inode_dotl(), so the next getattr will be read from cache again as
long as there is no other operations which invalidate the cache.

Change-Id: I69af51452c288d80c1f96c55402f7f99f9c2f7e4
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/#/c/photon/+/11803/
Tested-by: http://photon-jenkins.eng.vmware.com:8083/job/Gerrit_Photon/13580/
Reviewed-by: Srivatsa S. Bhat, Amod Mishra, Mounesh Badiger
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11803
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>