libsolv: Further extend choicerule filtering check
When we install tuned, instead of updating to latest
linux version libsolv was forcing to install linux-aws
package.

There is actually a countermeasure in the code
(the so-called "choice rules") to prevent the
installation of new packages if a dependency can also be
met by updating a package, but for some reason it
does not work in the case we install tuned.

This commit will fix this issue.

Change-Id: If01463e554138ad8a19d856ef0a2d57cb6f0b616
Signed-off-by: Tapas Kundu <tkundu@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/11832
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ankit Jain <ankitja@vmware.com>