gdb: Fix runtime issue with python3.9
We see a runtime crash when we try to start gdb
with python3.9.
we find that PyEval_ReleaseLock() has been deprecated.
It recommends using PyEval_SaveThread or PyEval_ReleaseThread()
instead.  In do_start_initialization, in gdb/python/python.c, we
can replace the calls to PyThreadState_Swap() and PyEval_ReleaseLock()
with a single call to PyEval_SaveThread.

Change-Id: I67285b52fc9efe24fae4ec484d7a0e01efbf507e
Signed-off-by: Tapas Kundu <tkundu@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12125
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ajay Kaher <akaher@vmware.com>