postgresql: Upgrade to 13.1 and fix set of issues
This addresses a set of issues found in the spec file of PostgreSQL:
- Fix list of BuildRequires.  While testing with rpmbuild -ba to test a
full build of the RPMs, I saw failures at various stages because of
packages missing.  Specifying the list of requirements from the start
eases the integration work.
- Update to 13.1 to address various bugs:
https://www.postgresql.org/docs/13/release-13-1.html
An annoying bug in the regression test suite, fixed by ee03baa caused
the test suite to fail across DST transitions.  This means that 13.0
would fail its %check phase half the year.
- Remove CVE-2016-5423.patch from the tree.  This referred to a patch
that is already integrated in the upstream code as of f0c7b789ab.

I have tested this patch after doing a full rpmbuild -ba validation.

Change-Id: Ibaeb510285b2060614d2653c6f97383034de50e4
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12334
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>