stalld: Support denylisting of tasks in stalld
This change allows certain tasks to be
specified from command line that will prevent
their priorities from being boosted even if
they are experiencing starvation.This is useful
in scenarios where boosting a certain application's
priority and in turn causing it to run interferes
with the purpose/logic of the application

The way to specify which task is to be ignored
(referred to here as the denylisting feature)
is through regular expressions. The regex of
the task name that needs to be ignored can be
passed in.

There are two ways to specify the denylisting.
Either threads can be specified to be ignored (-i)
or processes can be specified to be ignored (-I).
When a process is specified to be ignored, all
the threads that are created by that process will
also be ignored from consideration for priority
boosting.

Change-Id: I23b9d964b595dfb63691557ede580293035005c3
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12143
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>