calico-node: Fix calico-node not ready when using calico policy
When using photon calico-node as docker image in calico-policy.yaml,
The calico-node doesnt get in ready state and throws warning that
"confd/processor.go 143: Cannot load template resources: confdir
'/etc/calico/confd' does not exist."

This was because instead of copying whole calico folder inside /etc
we copy the content of it directly into /etc. This patch fixes it.

Change-Id: I719d40654739fe6ccf59762548c647c07d2c985c
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12398
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>