kernels: Fix /boot/photon.cfg symlink when /boot is a separate partition
If /boot is a separate partition (eg: if the system is setup that way
using custom partitioning during installation from ISO), then all the
paths in grub.cfg will be relative to that partition, which implies
that they won't have the /boot prefix in their path names. For
example, the menu entry will have /$photon_linux (instead of
/boot/$photon_linux), and photon.cfg will be referred to as
/photon.cfg.

On a system installed this way, if we uninstall a kernel package, the
photon.cfg symlink fixup script will point the (now) dangling
photon.cfg symlink back to an existing kernel's cfg, by hardcoding the
/boot/ prefix. That is, on a system with linux-esx installed, if we
run:

tdnf install linux -y
tdnf remove linux -y

Now, `readlink /boot/photon.cfg` shows:
/boot/linux-esx.cfg

instead of:
linux-esx.cfg

This is problematic because, after a reboot, grub2 won't be able to
find the path "/boot/linux-esx.cfg", because it is executing out of
the /boot partition itself, which is mounted at /. Hence, the file is
actually available at /linux-esx.cfg. Due to this mismatch, the OS
fails to boot with "error: not a regular file" message from grub2
(since the photon.cfg symlink cannot be resolved to the actual file it
points to).

Fix this issue in the %postun section by always pointing the
/boot/photon.cfg symlink to the appropriate linux-*.cfg, but relative
to the directory in which photon.cfg resides. (This is what the %post
section does too, which works well).

Change-Id: Ia892108cc892ed2b9615b77cfa0919423ffc5858
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/10606
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>