Redesign PostgreSQL RPMs for Photon
This commit reshapes the RPMs of PostgreSQL and splits things in a way
similar to RHEL and other major Linux distributions:
- postgresql for the client binaries, which requires postgresql-libs.
- postgresql-libs for the client libraries.
- postgresql-server for the server-side utilities, which requires
postgresql and postgresql-libs.
- postgresql-contrib for the contrib/ modules.
- postgresql-docs with the html documentation.
- postgresql-devel with all the elements needed to develop code related
to PostgreSQL.

While on it, this commit cleans up some compatibility links that
remained around after the major version bump to 10.  Anything using
the past name would be likely broken anyway as many option names have
changed at the same time as the binary names, so keeping these around
does not help much.

Change-Id: I306473b065212d5c3fef960b10eb2a55bd227aae
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12435
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>