perl-IPC-Run: Initial commit
IPC::Run is a perl module that can be used to run system() and
background processes with piping, redirections and ptys.  Getting this
module in Photon is required to expand regression tests of PostgreSQL,
that depend on it.

Change-Id: Ic330fcdd6463c3e33b44ced717e3507186d0655f
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12511
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>
Reviewed-by: Anish Swaminathan <anishs@vmware.com>
(cherry picked from commit fb1fa05c4966db2046df22d1dfa9cbe2465ed23b)
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12533
Tested-by: Anish Swaminathan <anishs@vmware.com>