postgresql: Add support for LLVM and JIT with --with-llvm
This adds a new package called postgresql-llvmjit that includes all the
JIT compilation code needed to be able to enable JIT in PostgreSQL via
LLVM.  This patch is the last one in the latest series to make
PostgreSQL on Photon equal to most major Linux distributions.

clang-devel requires the addition ncurses devel to allow the correct
fetching of -lcurses when compiling the JIT code for Postgres.

Change-Id: I05ed49d74f333c9940fd06420cdf94b034ca7874
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12635
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>