linux-esx: Introduce kernel panic on initramfs unpack failure for CRX
During the init sequence, when there is an error
in unpacking initramfs, the kernel was
only adding a log to dmesg and continued in the init
sequence. This eventually resulted in a panic being
thrown in the userspace init code(CRX) without any
indication of the reason why the error was
encountered.

Instead of allowing the kernel to continue with
init, add a panic upon encountering this error
to prevent any problems in userspace.

This change also consolidates all panics that are
specific to CRX under one kernel parameter -
panic_on_initramfs_error

Change-Id: I6b140d7ea11393311b5aa1c2c167ee89676a06c2
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12278
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>