postgresql12: Add PostgreSQL packages based on version 12
The main packages of Photon, named "postgresql", are based on the major
version 13.  This commit adds a new set of packages based on PostgreSQL
12 to be able to satisfy binary upgrade with pg_upgrade, that requires
the old and the new binaries and libraries to be able to work.  This new
set of packages can be installed in parallel of the main ones, meaning
that users can still choose to rely on a past version.  This also means
that the contents of the past version need to use a dedicated path,
chosen as /usr/pgsql/12 here, with the following structure:
- bin/ for all the binaries.
- lib/ for the libraries.
- include/ for the include files.
- doc/ for the man pages and HTML documentation.
- share/ for extension data, etc.

The same RPM layer as the existing "postgresql" packages is chosen for
ease of integration.  All the features of the main packages are also
supported here.

Change-Id: If2cc0c0265dd8d6b73371dc1e5b76370d131fd38
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/12665
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>