monitoring-plugins: Add the monitoring-plugins package to Photon
monitoring-plugins is a package that includes plugins to monitor
the state of hosts and the network. This package requires the perl
modules of net-snmp.

As an additional todo, the check_disk_smb plugin requires samba-client
and few other packages not added to Photon. This plugin may not work
as expected.

Change-Id: I985a77e878b3ed5531db70d8b6700463081dda93
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13504
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>