monitoring-plugins: Add the monitoring-plugins package to Photon
monitoring-plugins is a package that includes plugins to monitor
the state of hosts and the network. This package requires the perl
modules of net-snmp.

As an additional todo, the check_disk_smb plugin requires samba-client
and few other packages not added to Photon. This plugin may not work
as expected.

Change-Id: I0989a8c1bf2a36339ac12c161b4e6956e560cbb7
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13502
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreyas Ranjan Behera <shreyasb@vmware.com>