systemd: timedate - fix SetNtp: wrong PropertiesChanged value
https://github.com/systemd/systemd/issues/9672

 timedate: increment reference count of sd_bus_message

The commit 5d28074 introduces a
barrier to suppress calling context_update_ntp_status() multiple times.
However, it just stores the address of sd_bus_message object. So,
when an address is reused on the subsequent message, then the status
of NTP clients are not updated.

This makes the stored message object is referenced by the context
object. So, the subsequent message is on cirtainly different address.

 timedate: defer the property changed signal until job of starting/sto…

…pping NTP service is finished

Before this, the property changed signal is emitted immediately after
StartUnit/StopUnit method is called. So, the running state of the NTP
client service may not updated.
This makes the timing of emitting property changed signal is deferred
until job of starting/stopping NTP client service is completed.

Fixes systemd#9672.

Change-Id: I0bf091173cbee7d0464e99e0bda8ca7944abdbdb
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13505
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>