linux-rt: Enable nohz for idle=poll
If the tick timer is enabled in idle poll, it is possible that the
tick_sched_timer and wakeup timer goes off very close to each other.
This scenario can be very cosly in a virtualized environment.
There will be some impact on the average case for wakeup from idle

Change-Id: I54e1324ed4962d74982450b394cd2847ebd82b0a
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13558
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>