stalld: Add error handling for thread creation failure
stalld creates a separate monitoring thread for every CPU that needs
to be monitored for starvation (in aggressive mode) or every CPU that
is about to have starving tasks (in conservative mode).  However, in
both cases, the return value of pthread_create() is not checked for
errors, which means that if thread creation fails (due to temporary
resource constraints, for instance), then stalld will be unaware of
the problem and those corresponding CPUs will go unmonitored,
potentially leading to starvation and system hangs.

Fix this by checking the return value of pthread_create() and retry if
the error code was EAGAIN, and report the issue otherwise. Thread
creation failure is catastrophic enough that we need to abort stalld
in case of a persistent failure, because the threads also handle
priority-boosting for starved tasks (so, stalld cannot function without
them).

Reported-by: Him Kalyan Bordoloi <bordoloih@vmware.com>
Reported-by: Sharan Turlapati <sturlapati@vmware.com>
Signed-off-by: Him Kalyan Bordoloi <bordoloih@vmware.com>
[ Srivatsa: Added commit message ]
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>

Change-Id: I92743cd9dea77d848dbb1192ba53903573340534
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13641
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Sharan Turlapati <sturlapati@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>