monitoring-plugins: Add the monitoring-plugins package to Photon
monitoring-plugins is a package that includes plugins to monitor
the state of hosts and the network. This package requires the perl
modules of net-snmp.

As an additional todo, the check_disk_smb plugin requires samba-client
and few other packages not added to Photon. This plugin may not work
as expected.

Change-Id: I01276cd856c99b11c1322d8ac0c5b3684c820118
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13667
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreyas Ranjan Behera <shreyasb@vmware.com>