installer: Added --releasever to tdnf install
This option helps tdnf to point to repo from which rpms need to
be downloaded. In recent version of tdnf this parameter needs to
be set.

Added a new key as release_version in installer.py.
This key is set in imagebuilder.py for building cloud images and
ova. For iso this key is passed as a command line argument to
isoInstaller.py.

Change-Id: Id3e68d5ce7952ae46346fda4c6d7e46dcecba1fb
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13778
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>