curl: Patching for Fixing CVE-2021-22945/6/7
CVE-2021-22945
--------------
When sending data to an MQTT server, libcurl could in some circumstances
erroneously keep a pointer to an already freed memory area and both use that
again in a subsequent call to send data and also free it *again*.

CVE-2021-22946
--------------
A user can tell curl to **require** a successful upgrade to TLS when speaking
to an IMAP, POP3 or FTP server (`--ssl-reqd` on the command line or
`CURLOPT_USE_SSL` set to `CURLUSESSL_CONTROL` or `CURLUSESSL_ALL` with
libcurl). This requirement could be bypassed if the server would return a
properly crafted but perfectly legitimate response.

This flaw would then make curl silently continue its operations **without
TLS** contrary to the instructions and expectations, exposing possibly
sensitive data in clear text over the network.

CVE-2021-22947
--------------
When curl connects to an IMAP, POP3, SMTP or FTP server to exchange data
securely using STARTTLS to upgrade the connection to TLS level, the server can
still respond and send back multiple responses before the TLS upgrade. Such
multiple "pipelined" responses are cached by curl. curl would then upgrade to
TLS but not flush the in-queue of cached responses and instead use and trust
the responses it got *before* the TLS handshake as if they were authenticated.

Using this flaw, it allows a Man-In-The-Middle attacker to first inject the
fake responses, then pass-through the TLS traffic from the legitimate server
and trick curl into sending data back to the user thinking the attacker's
injected data comes from the TLS-protected server.

Change-Id: Ic05f6cfff488722253917d9642f5df4ff00d8a09
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14292
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>