postgresql: Upgrade to PostgreSQL 14 and add upgrade package for 13
This is an upgrade of the main set of PostgreSQL RPMs to include 14, the
latest version of upstream.  Like what has been done for 12 as of
7272162, this commit adds an upgrade package for 13, so as it becomes
possible to use pg_upgrade from this version to the latest one supported
in the main packages.  The packages of 12, 13 and 14 can be installed in
parallel as they use paths that do not conflict.

Change-Id: I4941b34ce55f04c7596716a220fd677b7d2a4dbd
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14390
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>