postgresql: Rework dependencies of -libs and -devel
Per discussion with Alexey Makhalov, the existing set of dependencies
listed in repmgr is misleading, as it has no need to directly depend on
libraries other than postgresql-devel for its build and postgresql-libs
at run time per the client nature of the command-line binary.

This commit reworks the set of dependencies in -libs and -devel so as we
have a complete listing of packages requires when installing any of
them, avoiding the need to pull in more dependencies in anything that
builds or links to PostgreSQL.  The -libs package needs to be able to
link to OpenSSL, OpenLDAP and krb5 at run time as libpq requires all of
them.  The -devel package, per its backend and frontend nature, has much
more requirements, as it needs to include most of the backend
dependencies like ICU, LLVM, etc (some headers include that).

This will help with a follow-up change to add --with-lz4 to the
PostgreSQL packages for LZ4 support, removing the need to change
repmgr.

Change-Id: Id7ab2ab683b18e433bd945fb51d553df9c876560
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14507
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>