postgresql: Add support for LZ4
This extends the capabilities of the backend to use LZ4 in its builds,
Postgres 14 adding the feature to use it for the compression of TOAST
data.  One dependency to lz4-devel is added for the build, and one
dependency to lz4 is added as a run-time dependency in the main
package per the backend requirements.

Change-Id: I1aee5386404ed5bab8355560e8cff533c2e9429c
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14478
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>