open-vm-tools: undo changes of 11.3.5-2 changes
If GOSC sub package doesn't depend on full version of open-vm-tools then
when a change is made in gosc sub package alone and version is bumped,
we might endup in a situation where only open-vm-tools is upgraded and
gosc sub package is not. So we must use %{release} as well in Requires.

Yes, the downgrade scenario will break but can be easily mitigated using
`tdnf downgrade -y open-vm-tools --allowerasing`.

Change-Id: Ib779bb446994004fe6530bf3061321d81e7bb9aa
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14567
Reviewed-by: Tapas Kundu <tkundu@vmware.com>
Tested-by: Tapas Kundu <tkundu@vmware.com>