trace-cmd: Add --poll option to trace-cmd
In the current design, trace-cmd reader processes need to be
woken up once data is available in the trace ring buffer. These
processes are woken up by ktimersoftd which itself is woken up
through self IPIs. On isolated CPUs, waking up ktimersoftd causes
the tick to turn on which affects latency.

When using trace-cmd for latency sensitive measurements like in RT,
we want to prevent these IPIs from happening. This change adds a --poll
option to trace-cmd record so that the reader processes poll for data
in the trace ring buffer.

This option needs to be used with caution since every trace-cmd process
will poll for data. Therefore, the system or HK CPUs need to have enough
bandwidth to carry out other tasks

Change-Id: I05f75d77468cfd05c1048b00d3868bde970f3967
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14616
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Keerthana K <keerthanak@vmware.com>