linux-esx: Added support for new filesystem tarfs
- TARFS file system to mount TAR archive.
- GNU/POSIX TAR Archive consists of header with information of the
files and directories. The start address of the file stored in a
field in the header itself which help to navigate to the file content.

- Tar archive can be attach to loop device and passed as input
that can be easily mounted instead of extraction using tarfs.
Tarfs does block device mount.
- TARFS only creates and maintains metadata, inodes and dentries
of files & directries. For a file, the entry for that particular
file will point to the start address offset of the TAR archive
whenever the file is read or executed.

- Mounting operation is very quick and avoids additional memory
allocation for file content data.

Change-Id: I0d82c6c62134f9b58179f5a45d22b1aed30e5cff
Signed-off-by: Ankit Jain <ankitja@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14642
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srinidhi Rao <srinidhir@vmware.com>