linux-secure: Fix fou_unknown patch for HCX
sysctl config "fou_key_multi_port" regsistration (using ipip_sysctl_init())
is done in ipip_init() which is the module load call of ipip. After this
registration on the same function ipip_init(), there is a check for
"fou_key_multi_port" variable to decide whether to register "fou_unknown_genl_family"
or not. This check should be removed here because "fou_key_multi_port"
is set to 0 by default in the code and gets enabled only through sysctl
which happens much later in the boot up sequence. Due to this check,
"fou_unknown_genl_family" registration never happens and causes a failure.

Change-Id: I82e2164a7f8dbb4f28982075dd604c5046353ced
Signed-off-by: Keerthana K <keerthanak@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/14704
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>