postgresql: Add upgrade package for PostgreSQL 13
The main RPM package set stored in postgresql.spec has been upgraded to
14, but there has been some asks for a fallback to 13 for some existing
users on Photon 4.0.  In the same spirit as the dev branch, this adds a
new package set for PostgreSQL 13, called "postgresql13".  This uses the
same base installation path, as of /usr/pgsql/13/, to allow multiple
major versions to be installed in parallel.

The RPM layer of "postgresql13" is the same as "postgresql" on 4.0, to
ease the switch to the new RPMs, requiring only a one-to-one rename.
The paths would still require an update to point to what's required,
likely with an addition of /usr/pgsql/13/bin/ to PATH.

Per reports from Yan Wang, Yang Jiao and Jiaolin Yang.

Change-Id: I3ed25fac9bcb0e9dc27a7dcfed1acd57c785943a
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15001
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>