linux-esx: ptp_vmw driver updates
Extend ptp_vmw driver's functionality to allow a VM to make time
adjustments to the PHC device, enabling mechanisms that allow
time sync agents running in a VM to provide time to other VMs
and the underlying host. Also add a module parameter probe_hv_port
that allows ptp_vmw driver to be loaded even when ACPI is disabled.

Change-Id: Iad30ba8a007d03fed3c53bf884db443fb0d49526
Signed-off-by: Deep Shah <sdeep@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15082
Reviewed-by: Vivek Thampi <vithampi@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>