util-linux: Fix for CVE-2021-3995 and CVE-2021-3996
CVE-2021-3996 and CVE-2021-3995 are side effects of: 
https://github.com/util-linux/util-linux/\
commit/5fea669e9ef0a08804f72bb40f859f239f68c30a.

CVE-2021-3995:
-------------
There is a vulnerability in the code which allows an unprivileged
user to unmount the FUSE filesystems that belong to certain other users;
for example, if our own uid is 1000, then we can unmount the FUSE
filesystems of the users whose uid is 100, 10, or 1.

CVE-2021-3996:
-------------
This vulnerability allows an unprivileged user to unmount other users'
filesystems that are either world-writable themselves (like /tmp) or
mounted in a world-writable directory.

For example, on Fedora, /tmp is a tmpfs, so we can mount a basic FUSE
filesystem named "/tmp/ (deleted)" (with FUSE's "hello world" program,
./hello) and unmount /tmp itself (a denial of service):

Change-Id: I2e82b22cdab27f816411953f97ff0fba6bd3fcf8
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15110
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Ajay Kaher <akaher@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>