openssl: Add jitterentropy provider
FIPS provider does not include an entropy source. And FIPS provider
was certified with entropy source in passive mode. That means, entropy
for FIPS crypto has to be provided froom outside of the FIPS boundary.

We have an entropy source in the kernel which was FIPS certified -
jitterentropy library. This entropy source can be obtained by the user
space through AF_ALG socket type.
https://www.kernel.org/doc/html/v4.11/crypto/userspace-if.html

Jitterentropy provider, that I'm adding with this commit, is a proxy
allowing OpenSSL FIPS crypto to consume entropy source from the kernel.

Both fips.so and jitterentropy.so providers must be loaded and activated
in order to switch OpenSSL to FIPS enabled mode.

Extras:
- post/postun logic of openssl-fips-provider package to enable/disable FIPS
  was redesigned to use corresponding provider config file in main OpenSSL
  config file. It will allow us to preserve user specific configuration in
  openssl.cfg during OpenSSL RPM update.
- external %{?with_fips} macro was removed and our build infra does not
  control enabling of FIPS build of OpenSSL as there is no more OpenSSL
  canister. Use defined in .spec file %{?fips} macro instead.

Change-Id: Id4d3a0f4badcb666653716206d5bee73a38dae91
Signed-off-by: Alexey Makhalov <amakhalov@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15144