bindutils: Version upgraded to 9.16.27
The package version is upgraded to 9.16.27 to address following CVEs -
- CVE-2021-25220
- CVE-2022-0396

Do not delete user and group unless uninstalled

As postun for the older pkg runs later than pre for
latest package, so postun was not checking if there
is an uninstall happening or upgrade/downgrade.
postun for the previous package was removing the
user and group irrespective of the upgrade/downgrade/uninstall
causing this issue.

Fix:
As we cannot fix the postun for the published pkgs, we need
to handle creating user/group in posttrans and fixed the
postun logic.

Change-Id: Ic00f4766b17d55c646bcfde1cf8669625c1bb329
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15441
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>