bindutils: do not delete user and group unless uninstalled
As postun for the older pkg runs later than pre for
latest package, so postun was not checking if there
is an uninstall happening or upgrade/downgrade.
postun for the previous package was removing the
user and group irrespective of the upgrade/downgrade/uninstall
causing this issue.

Fix:
As we cannot fix the postun for the published pkgs, we need
to handle creating user/group in posttrans and fixed the
postun logic.

Change-Id: I598cc45d505ef0867f9bd29a872023f4951ae08e
Signed-off-by: Tapas Kundu <tkundu@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/15514
Tested-by: gerrit-photon <photon-checkins@vmware.com>