linux-rt: Backport hrtick changes to fix lost timer wakeups
HRTICK code in the scheduler is vulnerable to a race condition
that can corrupt the hrtimer rbtree. The corruption can result in
sleeping tasks not waking up after their timer expiry. Backport
changes from upstream to 4.19 to fix the race condition. While at it,
bring in all other changes related to HRTICK from upstream.

The patchset includes the following changes -
1) 0001-sched-Mark-hrtimers-to-expire-in-hard-interrupt-cont.patch -
   Only a cosmetic change that marks timer interrupts as executing in hard
   irq context.

2) 0002-sched-features-Fix-hrtick-reprogramming.patch -
   Fixes a race condition in hrtick's hrtimer reprogramming

3) 0003-sched-features-Distinguish-between-NORMAL-and-DEADLI.patch -
   Distinguishes between the features for enabling hrtick
   for CFS (HRTICK) and for SCHED_DEADLINE(HRTICK_DL)

Change-Id: Id19563ba596c821978ae6b89daf73f3b72d2267d
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/16047
Tested-by: Sharan Turlapati <sturlapati@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>